Colocate all leprecondo functionality in LeprecondoManager, fix updated handling of plurals and vowel-starts in discovery messages, parse installed furniture when out of rearrangements
