+++ src/net/sourceforge/kolmafia/request/FightRequest.java
@@ -14,12 +14,9 @@ import java.util.TimeZone;
 import java.util.TreeMap;
 import java.util.TreeSet;
 import java.util.function.Function;
-import java.util.function.Predicate;
 import java.util.regex.Matcher;
 import java.util.regex.Pattern;
 import java.util.stream.Collectors;
-import java.util.stream.IntStream;
-import java.util.stream.Stream;
 import net.sourceforge.kolmafia.AdventureResult;
 import net.sourceforge.kolmafia.AdventureResult.AdventureLongCountResult;
 import net.sourceforge.kolmafia.AreaCombatData;
@@ -100,6 +97,7 @@ import net.sourceforge.kolmafia.session.GrimstoneManager;
 import net.sourceforge.kolmafia.session.InventoryManager;
 import net.sourceforge.kolmafia.session.IslandManager;
 import net.sourceforge.kolmafia.session.JuneCleaverManager;
+import net.sourceforge.kolmafia.session.LeprecondoManager;
 import net.sourceforge.kolmafia.session.LimitMode;
 import net.sourceforge.kolmafia.session.LocketManager;
 import net.sourceforge.kolmafia.session.LoginManager;
@@ -7754,7 +7752,7 @@ public class FightRequest extends GenericRequest {
     }
 
     // Leprecondo actions are tagged this way
-    if (FightRequest.handleLeprecondo(str, image)) {
+    if (LeprecondoManager.handlePostCombatMessage(str, image)) {
       return;
     }
 
@@ -8676,132 +8674,6 @@ public class FightRequest extends GenericRequest {
     }
 
     return false;
-  }
-
-  private static final String[] LEPRECONDO_FURNITURE =
-      new String[] {
-        "buckets of concrete",
-        "thrift store oil painting",
-        "boxes of old comic books",
-        "second-hand hot plate",
-        "free mattress",
-        "beer cooler",
-        "UltraDance karaoke machine",
-        "cupcake treadmill",
-        "internet-connected laptop",
-        "sous vide laboratory",
-        "whiskeybed",
-        "complete classics library",
-        "ultimate retro game console",
-        "four-poster bed",
-      };
-
-  private static final Map<String, String> LEPRECONDO_FULFILMENT_TO_NEED =
-      Map.ofEntries(
-          Map.entry("spends a few minutes swinging around his buckets of concrete", "exercise"),
-          Map.entry(
-              "wanders over and spends some time studying his painting", "mental stimulation"),
-          Map.entry("plops down and reads some old comics", "dumb entertainment"),
-          Map.entry("fires up his second-hand hot plate", "food"),
-          Map.entry("heads to his filthy mattress for a nap", "sleep"),
-          Map.entry("grabs a beer from the beer cooler", "booze"),
-          Map.entry(
-              "wanders over to the karaoke machines and sings a few popular songs",
-              "dumb entertainment"),
-          Map.entry(
-              "sets his karaoke machine to dance mode and careens around the room", "exercise"),
-          Map.entry("spends a while running on his cupcake treadmill", "exericse"),
-          Map.entry("grabs a cupcake from his treadmill, and tosses you one as well", "food"),
-          Map.entry(
-              "plays beer pong for a while, but stops drinking before he finishes the whole table's worth",
-              "booze"),
-          Map.entry(
-              "spends some time arguing with people about politics on social media",
-              "dumb entertainment"),
-          Map.entry(
-              "watches some home improvement videos and takes some notes", "mental stimulation"),
-          Map.entry(
-              "whips up a giant meal in the sous vide machine and shares it with you", "food"),
-          Map.entry("reads the manual for his sous vide machine", "mental stimulation"),
-          Map.entry("extracts some of the whiskey from his bed", "booze"),
-          Map.entry("takes a long nap in his whiskeybed", "sleep"),
-          Map.entry("reads some books about military history", "mental stimulation"),
-          Map.entry("play old video games for a few hours", "dumb entertainment"),
-          Map.entry("takes a long nap in his nice bed", "sleep"));
-
-  static final Pattern LEPRECONDO_FURNITURE_DISCOVERY =
-      Pattern.compile("spots a (.*?) and runs out of his condo\\.  He drags it back");
-
-  static final Pattern LEPRECONDO_NEED = Pattern.compile("is upset that his (.*?) need wasn't met");
-
-  private static boolean handleLeprecondo(final String text, final String image) {
-    if (!image.equals("familiar2.gif")) return false;
-
-    // Check for new furniture discovery
-    var discovery = LEPRECONDO_FURNITURE_DISCOVERY.matcher(text);
-    if (discovery.find()) {
-      var discovered =
-          IntStream.range(1, LEPRECONDO_FURNITURE.length)
-              .filter((i) -> discovery.group(1).startsWith(LEPRECONDO_FURNITURE[i - 1]))
-              .findAny()
-              .orElse(0);
-
-      if (discovered == 0) return false;
-
-      Preferences.setString(
-          "leprecondoDiscovered",
-          Stream.concat(
-                  Arrays.stream(Preferences.getString("leprecondoDiscovered").split(","))
-                      .filter(Predicate.not(String::isBlank))
-                      .map(StringUtilities::parseInt),
-                  Stream.of(discovered))
-              .sorted()
-              .distinct()
-              .map(String::valueOf)
-              .collect(Collectors.joining(",")));
-      return true;
-    }
-
-    // Check for unmet need
-    var unmetNeed = LEPRECONDO_NEED.matcher(text);
-    if (unmetNeed.find()) {
-      processLeprecondoNeed(unmetNeed.group(1));
-      return true;
-    }
-
-    // Check for met need
-    var metNeed =
-        LEPRECONDO_FULFILMENT_TO_NEED.entrySet().stream()
-            .filter(e -> text.contains(e.getKey()))
-            .map(Map.Entry::getValue)
-            .findAny()
-            .orElse(null);
-    if (metNeed != null) {
-      processLeprecondoNeed(metNeed);
-      return true;
-    }
-
-    return false;
-  }
-
-  protected static void processLeprecondoNeed(final String need) {
-    if (Preferences.getString("leprecondoCurrentNeed").equals(need)) return;
-    Preferences.setString("leprecondoCurrentNeed", need);
-    Preferences.setInteger("leprecondoLastNeedChange", KoLCharacter.getCurrentRun());
-    var history = Preferences.getString("leprecondoNeedOrder");
-
-    // Already seen this need
-    if (history.contains(need)) {
-      // If we do not have a complete set, the data is broken. Start counting again
-      if (history.split(",").length != 6) {
-        Preferences.setString("leprecondoNeedOrder", need);
-      }
-
-      return;
-    }
-
-    Preferences.setString(
-        "leprecondoNeedOrder", String.join(",", history) + ((history.isEmpty()) ? "" : ",") + need);
   }
 
   private static void logSkillAcquisition(String skillName, final TagStatus status) {
