+++ test/net/sourceforge/kolmafia/request/FightRequestTest.java
@@ -3970,29 +3970,5 @@ public class FightRequestTest {
         assertThat("leprecondoLastNeedChange", isSetTo(42));
       }
     }
-
-    @ParameterizedTest
-    @CsvSource(
-        value = {
-          "booze|''|booze",
-          "food|sleep,exercise,mental stimulation|sleep,exercise,mental stimulation,food",
-          "food|sleep,food,booze,exercise|food",
-          "food|sleep,mental stimulation,food,dumb entertainment,booze,exercise,booze|food",
-          "exercise|sleep,mental stimulation,food,dumb entertainment,exercise,booze|sleep,mental stimulation,food,dumb entertainment,exercise,booze",
-        },
-        delimiter = '|')
-    void calculatesNeedOrder(final String need, final String order, final String expected) {
-      var cleanups =
-          new Cleanups(
-              withItem("Leprecondo"),
-              withProperty("leprecondoDiscovered"),
-              withProperty("leprecondoCurrentNeed", ""),
-              withProperty("leprecondoNeedOrder", order),
-              withFight(0));
-      try (cleanups) {
-        FightRequest.processLeprecondoNeed(need);
-        assertThat("leprecondoNeedOrder", isSetTo(expected));
-      }
-    }
   }
 }
